requirements.txt
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
einops==0.8.0
encodec==0.1.1
filelock==3.16.1
fsspec==2024.10.0
funcy==2.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
packaging==24.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
s3transfer==0.10.3
safetensors==0.4.5
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
suno-bark==0.1.5
sympy==1.13.1
tokenizers==0.20.2
torch==2.5.1
torchaudio==2.5.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
